+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -279,6 +279,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`hooks-transaction-hash-buf-len
+  <clang-tidy/checks/hooks-transaction-hash-buf-len>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`hooks-validate-buf-len
   <clang-tidy/checks/hooks-validate-buf-len>` check.